Create movie model. Set up movie seed data.
The movie model uses ActiveRecord as an ORM to map models to database
table names. Movie(model) -> movies(database table)

Self Check
1. Why are `where` and `find` class methods rather than instance method?

Instance methods operate on one instance of the class, but until we look
up one or more objects, we have no instances to operate on.

2. Do Rails models acquire the methods `where` and `find` via (a)
inheritance, or (b) mix-in?

They inherit from the ActiveRecord::Base.

NOTE:
ActiveRecord uses convention over configuration to infer databse table
names from the names on model classes, and to infer the names and types
of the columns (attributes) associated with a given kind of model.