Complete movies controllers and associated views.
Self Check
1. Why don't the controller methods for the `new` action (new_movie_path)
and `create` action (movies_path) take any arguements when the `show`
and `update` helpers do?

Show and update operate on existing movies, so they take an argument to
identify which movie to operation on. New and fcreate by definition
operate on not-yet-existing movies.

2. Why doesn't the helper method for the Index action take an argument?

Index just shows a list of all the movies, so no argument is needed to
distinguish which movie to operate on.

3. Why is there no end corresponding to the `do` in haml templates?

Unlike Ruby itself, Haml relies on indentation to indicate nesting, so
the end is supplied by Haml when executing the Ruby code in the `do`.